fix race condition with max_connect_errors and failed connection attempts
The max_connect_errors count increases each time a failed connection attempt
happens.  These connection attempts appear to be happening from some the
of the services coming up, before the service has been configured in
mysql.

If the default limit of 10 is hit before any of the services can be
configured in mysql, then all subsequent connection attempts (including
those that are attempts to configure the missing users, which would fix
the problem) fail and increment this count.  There is no way back after
this point.

This commit raises the default limit high enough so that the services
can all be configured before the limit is hit.  This stops the problem
in it's tracks.