do not enable binlogging whilst install is still happening
by setting an attribute for 'log_bin', we make the mysql recipe drop
configuration that enables binlogging from the start.  Sometimes there
is a race condition wherby the config file gets dropped in, and mysql is
restarted, before the mysql_install_db script has finished.  This causes
some of the setup statements (create mysql database, performance_schema
database etc) to be passed into the binlog.  This is BAD, since when
replication starts from a second node, it tries to run these staements
against itself - it has already run these statements as part of it's own
install, and chaos and death ensues.

Due to the way chef does not allow you to modify an attribute part way
through a recipe, we are going to have to enable binlogging in a
slightly different way - by fropping a conf.d config file just for
binlogging, later on, at a time that is under our control.